% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.cor.r
\name{af.cor}
\alias{af.cor}
\title{Function af.cor}
\usage{
af.cor(dat, wei = NULL, use = "pairwise.complete.obs")
}
\arguments{
\item{data}{Numeric data.frame.}
}
\description{
Provides (unweighted or weighted) correlation matrix between variables.
}
\details{
Provides (unweighted or weighted) correlation matrix between variables.
}
\examples{
af.cor(dat=sapply(1:10,function(x)rnorm(100)));
}
\keyword{modeling}
