% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.ctm.r
\name{af.ctm}
\alias{af.ctm}
\title{Function af.ctm}
\usage{
af.ctm(data = c("Hallo Welt", "Hello World"), k = 3, method = "VEM",
  nstart = 10, seed = 0)
}
\description{
Function af.ctm
}
\keyword{modeling}
