% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.fa.r
\name{af.fa}
\alias{af.fa}
\title{Function af.fa}
\usage{
af.fa(data, nfactors = NULL, rotate = "varimax", fm = "pa",
  plot = T, ...)
}
\arguments{
\item{data}{Numeric data.frame.}
}
\description{
Factor analysis of data.frame.
}
\details{
Factor analysis of data.frame.
}
\examples{
af.fa(data.frame(x=rnorm(100),y=rnorm(100)+scale(1:100),z=rnorm(100)+scale(1:100)));plotFA(fa)
}
\keyword{modeling}
