% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.fit.r
\name{af.fit}
\alias{af.fit}
\title{Function af.fit}
\usage{
af.fit(data.train, output = NULL, rep.nnet = 10)
}
\arguments{
\item{data.train}{data.frame containing training data.}

\item{output}{index or name of criterion in training data.}

\item{rep.nnet}{number of random initialisations that should be applied for choosing af.nnet.}
}
\description{
Fits nnet and linear regression (with and without interaction terms) to a data.frame.
}
\examples{
data=data.frame(y=rnorm(100),x1=rnorm(100),x2=rnorm(100));
fit.quant(data,"y1")
}
\keyword{modeling}
