% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.hclust.r
\name{af.hclust}
\alias{af.hclust}
\title{Function af.hclust}
\usage{
af.hclust(
  data,
  dist.function = function(x) as.dist(1 - cor(x)),
  method = "complete",
  k = NULL,
  plot = T
)
}
\arguments{
\item{data}{Numeric data.frame.}
}
\description{
Cluster analysis of data.frame.
}
\details{
Cluster analysis of data.frame.
}
\examples{
af.hclust(data.frame(x=rnorm(100),y=rnorm(100)+scale(1:100),z=rnorm(100)+scale(1:100)));
}
\keyword{modeling}
