% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.lasso.r
\name{af.lasso}
\alias{af.lasso}
\title{Function af.lasso}
\usage{
af.lasso(data, output = NULL, control = caret::trainControl(method =
  "boot", number = 30))
}
\arguments{
\item{data}{data.frame containing data.}

\item{output}{index or name of criterion in data.}

\item{y}{numeric vector.}
}
\description{
Applies LASSO regression and returns data.frame with important predictors.
}
\examples{
data=data.frame(y=rnorm(100),x1=rnorm(100),x2=rnorm(100));
af.lasso(data,"y1")
}
\keyword{preprocessing}
