% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.model.r
\name{af.model}
\alias{af.model}
\title{Function af.model}
\usage{
af.model(formula, data, direction = "top", model = lm, scope = NULL,
  ...)
}
\arguments{
\item{formula}{A formula specifying the criterion and its relation to all predictors possibly relevant.}

\item{data}{Numeric data.frame.}

\item{direction}{Character element specifying the stepwise procedure to be applied (besides "forward", "backward" or "both", possible values are "none", "adaptive", "complete" or "top". It's also possible to add a number after "top" (e.g. "top 3") in order to specify the maximum number of predictors to be examined). Defaults to "top".}

\item{model}{Function of an approproate class. Defaults to lm.}
}
\description{
Statistical model based on the most relevant predictors as chosen using a stepwise procedure.
}
\details{
This function fits a statistical model based on the most relevant predictors as chosen using a stepwise procedure.
}
\examples{
af.hclust(data.frame(x=rnorm(100),y=rnorm(100)+scale(1:100),z=rnorm(100)+scale(1:100)));
}
\keyword{modeling}
