% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.nnet.r
\name{af.nnet}
\alias{af.nnet}
\title{Function af.nnet}
\usage{
af.nnet(data.train, output = NULL, rep.nnet = 10, seed = 0,
  plot = F, size = NULL, decay = 0, linout = T, trace = F,
  maxit = 1000, preProc = c("center", "scale"),
  control = caret::trainControl(method = "boot", number = 10), ...)
}
\arguments{
\item{data.train}{data.frame containing training data.}

\item{output}{index or name of criterion in training data.}

\item{size}{numeric value determining the number of hidden units.}

\item{decay}{numeric value determining the decay parameter.}
}
\description{
Fits multiple initializations of nnet to scaled data and returns best nnet-object.
}
\details{
Fits multiple initializations of nnet to scaled data and returns best nnet-object. If size or decay are NULL, 10-fold cross validation is applied to find an optimal set of parameters
}
\examples{
 
}
\keyword{modeling}
