% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.sensitivity.r
\name{af.sensitivity}
\alias{af.sensitivity}
\title{Function af.sensitivity}
\usage{
af.sensitivity(model, x = NULL, y = NULL, steps = 100,
  splitseq = seq(0, 1, by = 0.25), plot = T, verbose = T, ...)
}
\arguments{
\item{model}{Any object that works as input for function predict.}

\item{x}{data.frame with numeric predictor variables.}

\item{y}{numeric criterion variable.}
}
\description{
Plots Lek-Pofiles and Fischer's Deltas.
}
\details{
Plots Lek-Pofiles and Fischer's Deltas (i.e., the difference in R^2 that results from omitting a predictor for each predictor, as described by Fischer(2015)).
}
\examples{
data=data.frame(y=rnorm(100)+scale(1:100),x1=rnorm(100)+scale(1:100),x2=rnorm(100));
l=lm(y~.,data=data)
sensitivity(l,data[c("x1","x2")],data["y"])
}
\keyword{modeling}
