% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.stm.r
\name{af.stm}
\alias{af.stm}
\title{Function af.stm}
\usage{
af.stm(data = c("Hallo Welt", "Hello World"), K = 3,
  lower.thresh = 0, removestopwords = F, ...)
}
\description{
Function af.stm
}
\keyword{modeling}
