% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.r
\name{bp}
\alias{bp}
\title{Function bp}
\usage{
bp(x = NULL, sd = NULL, cex = 0.7, beside = T, horiz = F,
  add.numbers = F, plot = T, main = NULL, xlim.factor = 1.5,
  xlim = NULL, names.arg = NULL, legend.text = NULL, col = NULL,
  ...)
}
\arguments{
\item{x}{Numeric vector, matrix or data.frame containing the values to be displayed.}

\item{sd}{Numeric vector, matrix or data.frame of same format as x containing standard deviations.}

\item{cex}{Size of fonts. Defaults to .7.}

\item{beside}{Logical value indicating if bars should be placed next to each other. Defaults to T.}

\item{horiz}{Logical value indicating if bars should be placed horizontal. Defaults to F.}

\item{add.numbers}{Logical value indicating if numbers should be placed above bars. Defaults to Fs.}

\item{plot}{Logical value indicating whether to plot the barplot. Defaults to T.}

\item{main}{Character vector with one element containing the barplot's title. Defaults to NULL}

\item{xlim.factor}{Numeric value for adding extra space to the right of the barplot. Defaults to 1.}

\item{xlim}{Numeric vector containing xlim.}

\item{names.arg}{Character vector containing names of bars. If NULL (default) colnames of x will be applied as names.arg.}

\item{legend.text}{Legend text. If NULL (default) rownames of x will be applied as legend.text.}

\item{col}{Vector containing the color of bars. If NULL (default) colors are generated based on the rainbow-palette.}

\item{...}{Additional graphical parameters for barplot.}
}
\description{
Custom barplot with labeled bars.
}
\details{
Plots a bar and adds labels and numbers to bars. Optionally allows for plotting standard deviations around each bar.
}
\examples{
bp(data.frame(group1=c(variable1=1,variable2=2),group2=c(variable1=3,variable2=4)),horiz=T,main="Example")
}
\keyword{plotting}
