% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.r
\name{classify}
\alias{classify}
\title{Function classify}
\usage{
classify(dat, cat = character(0), cod = character(0),
  pos = numeric(0), resort = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Function classify
}
\keyword{mining}
\keyword{text}
