% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterWordFrequencies.r
\name{clusterWordFrequencies}
\alias{clusterWordFrequencies}
\title{Function clusterWordFrequencies}
\usage{
clusterWordFrequencies(corpus, cor = F, binary = T, min = 0,
  plot = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Cluster analysis of words based on a character vector.
}
\details{
Cluster analysis of words based on a character vector.
}
\examples{
clusterWordFrequencies(c("<1>hello</1> world"),c("<1>hello</1> world","<2>hallo</2>"))
}
\keyword{mining}
\keyword{text}
