% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeByPassage.r
\name{codeByPassage}
\alias{codeByPassage}
\title{Function codeByPassage}
\usage{
codeByPassage(corpus, passage, coding, context = 0, mess = T,
  assign = T, subcorpus = NULL)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Adds code tags to certain passages of a character vector.
}
\details{
Adds code tags to certain passages of a character vector.
}
\examples{
codeByPassage("hello world hello world hello world hello world","hello","code1")
}
\keyword{mining}
\keyword{text}
