% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeByPosition.r
\name{codeByPosition}
\alias{codeByPosition}
\title{Function codeByPosition}
\usage{
codeByPosition(corpus, position, coding, mess = T, assign = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Adds code tags to certain positions of a character vector.
}
\details{
Adds code tags to certain positions of a character vector.
}
\examples{
codeByPosition("hello world hello world hello world hello world",data.frame(c(1,1,5)),"code1")
}
\keyword{mining}
\keyword{text}
