% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeByRegex.r
\name{codeByRegex}
\alias{codeByRegex}
\title{Function codeByRegex}
\usage{
codeByRegex(corpus, passage, coding, context = 0, mess = T,
  assign = T, subcorpus = NULL)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Adds code tags to certain regex-passages of a character vector.
}
\details{
Adds code tags to certain regex-passages of a character vector.
}
\examples{
codeByRegex("hello world hello world hello world hello world","hello","code1")
}
\keyword{mining}
\keyword{text}
