% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherenceLDA.r
\name{coherenceLDA}
\alias{coherenceLDA}
\title{Function coherenceLDA}
\usage{
coherenceLDA(lda, dtm, n = 0)
}
\arguments{
\item{lda}{Object of class LDA.}

\item{dtm}{document-term matrix.}

\item{n}{Number of terms for determining document-coocurence. If 0 (default) n is set to the number of terms in the dtm.}
}
\description{
Calculates semantic coherence of an LDA.
}
\keyword{modeling}
