% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.r
\name{cols}
\alias{cols}
\title{Function cols}
\usage{
cols(num = 1, col2 = NULL, col1 = NULL)
}
\arguments{
\item{num}{Number of colors to return.}

\item{col2}{Character vector specifying the darker color of the specturm to be returned.}

\item{col1}{Character vector specifying the lighter color of the specturm to be returned.}
}
\description{
Returns a vector of colors of a certain length.
}
\details{
Returns a vector of colors of a certain length. If col1 and col2 are NULL (default) a rainbow palette is returned. If col2 is specified, col1 is set to "white". If col1 is specified, col2 is set to "black". In addition to regular color labels, the cols-function accepts some pre-defined color-labels (such as "jddm" or "jddmLight" for the Logo-colors of the Journal of Dynamic Decision Making)
}
\examples{
bp(1:10,col=cols(10,"jddm"))
}
\keyword{plotting}
