% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols2.r
\name{cols2}
\alias{cols2}
\title{Function cols2}
\usage{
cols2(col, transparency = 0.7)
}
\arguments{
\item{col}{Character vector specifying the colors to be transformed.}

\item{transparency}{Numberic value specifying the transparency of the colors to be returned. Defaults to .3}
}
\description{
Returns a vector of transparent colors based on a vector of colors.
}
\details{
Returns a vector of transparent colors based on a vector of colors.
}
\examples{
cols2(cols(10,"jddm")))
}
\keyword{plotting}
