% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDocs.r
\name{compareDocs}
\alias{compareDocs}
\title{Function compareDocs}
\usage{
compareDocs(a = NULL, b = NULL, legendtext = c("a", "b"),
  relative = T, ...)
}
\arguments{
\item{a}{Character value.}

\item{b}{Character value.}

\item{relative}{Logical value. Defaults to T.}

\item{main}{Character vector with one element containing the plot's title. Defaults to NULL}

\item{show.text}{Logical value indicating whether bubbles should be labelled with names(vec). Defaults to T.}

\item{beak.names}{Logical value indicating whether line breaks should be added after each whitespace. Defaults to F.}

\item{beak.names}{Logical value indicating whether values of vec should be added to vec's names before labelling the bubbles.}

\item{cex}{Numeric value indicating the text size. Defaults to .8.}

\item{a}{Numeric value multiplied with normalizd vec. Defaults to 90.}

\item{b}{Numeric value added to normalized vec*a. Defaults to 10.}

\item{col}{Character vector indicating the bubbles' color. If NULL (default) the rainbow palette is applied to vec's ranks.}
}
\description{
Plots difference between two documents as packed bubble chart.
}
\details{
Plots vector as packed bubble chart. Returns coordinates and radius of each bubble
}
\examples{
compareDocs("This is document 1","this is another document")
}
\keyword{plotting}
