% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlateCodeFrequencies.r
\name{correlateCodeFrequencies}
\alias{correlateCodeFrequencies}
\title{Function correlateCodeFrequencies}
\usage{
correlateCodeFrequencies(corpus, binary = T, min = NULL,
  returnVector = T, plot = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Correlates code frequencies across documents.
}
\details{
Correlates code frequencies across documents.
}
\examples{
correlateCodeFrequencies(c("<1>hello</1> world","<1>hello</1> world","<2>hallo</2>"))
}
\keyword{mining}
\keyword{text}
