% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlateWordFrequencies.r
\name{correlateWordFrequencies}
\alias{correlateWordFrequencies}
\title{Function correlateCodeFrequencies}
\usage{
correlateWordFrequencies(
  corpus,
  binary = T,
  min = NULL,
  returnVector = T,
  plot = T
)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Correlates word frequencies across documents.
}
\details{
Correlates word frequencies across documents.
}
\examples{
correlateWordFrequencies(c("<1>hello</1> world","<1>hello</1> world","<2>hallo</2>"))
}
\keyword{mining}
\keyword{text}
