% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.r
\name{correlationplot}
\alias{correlationplot}
\title{Function correlationplot}
\usage{
correlationplot(x = NULL, stars = T, numbers = T, cex = 0.6,
  show.legend = F, ...)
}
\arguments{
\item{x}{Numeric data.frame or matrix containing the vectors whose correlations should be displayed.}

\item{stars}{Logical value indicating whether significant results should be marked with stars. Defaults to T.}

\item{numbers}{Logical value indicating whether correlations should be plotted numerically. Defaults to T.}

\item{cex}{Numerical value specifying the relative size of texts to be displayed. Defaults to .6.}

\item{show.legend}{Logical value indicating whether a legend shoud be displayed. Defaults to F.}

\item{...}{Additional arguments to be passed to psych::cor.plot.}
}
\description{
Plots correlations based on a data.frame.
}
\details{
Plots correlations based on a data.frame.
}
\examples{
set.seed(0);correlationPlot(data.frame(x=rnorm(100),y=rnorm(100),z=rnorm(100)))
}
\keyword{plotting}
