% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decollide.r
\name{decollide}
\alias{decollide}
\title{Function decollide}
\usage{
decollide(
  x,
  y,
  text,
  jitter = F,
  cex = 1,
  font = 1,
  frame = 1.5,
  plot = F,
  lock.x = F,
  lock.y = F,
  verbose = F,
  repetitions = 100
)
}
\arguments{
\item{x}{Numeric vector containing x-coordinates of text elements.}

\item{y}{Numeric vector containing x-coordinates of text elements.}

\item{text}{Character vector containing text-elements to be plotted.}

\item{jitter}{Logical value specifying whether to add a small random component after each step.}

\item{cex}{Numeric value specifying the maximum number of steps. Defaults to 100.}

\item{frame}{Numeric value specifying the relative font size of an invisible frame around the text. Defaults to 1.5}

\item{lock.x}{Logical value specifying whether to skip shifts on the x-axis. Defaults to F.}

\item{lock.y}{Logical value specifying whether to skip shifts on the y-axis. Defaults to F.}

\item{verbose}{Logical value specifying whether to give information on the process after each step.}
}
\description{
Returns a matrix with coordinates and text for plotting texts without collisions.
}
\details{
Returns a matrix with coordinates and text for plotting texts without collisions. Iteratively shifts text elements until all collisions are eliminated or the maximum number of repetitions is reached.
}
\examples{
set.seed(0);decollide(x=rnorm(10),y=rnorm(10),text=paste0(1:10,":",round(rnorm(10),3)),cex=1,font=1,plot=T)
}
\keyword{plotting}
