% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltas.r
\name{deltas}
\alias{deltas}
\title{Function deltas}
\usage{
deltas(
  model,
  x = NULL,
  y = NULL,
  fun = mean,
  plot = T,
  main = "Fischer's Delta",
  xlab = "Predictor",
  ylab = expression(R^2),
  col = NULL,
  ...
)
}
\arguments{
\item{model}{Any object that works as input for function predict.}

\item{x}{data.frame with numeric vectors.}

\item{y}{numeric vector.}

\item{fun}{A function that specifies what a predictor should be replaced with to remove its influence. Defaults at mean.}

\item{plot}{Logical value indicating whether to plot the barplot. Defaults zo T.}

\item{main}{Character vector with one element containing the barplot's title. Defaults to "Fischer's Delta".}

\item{xlab}{Character vector with one element containing the barplot's x-axis label. Defaults to "Predictor".}

\item{ylab}{Character vector with one element containing the barplot's y-axis label. Defaults to "\eqn{R^2}".}

\item{col}{Vector containing the color of bars. If NULL (default) colors are generated based on the rainbow-palette.}

\item{...}{Additional graphical parameters for barplot.}
}
\description{
Calculates the difference in \eqn{expression(R^2)} that results from omitting a predictor for each predictor.
}
\details{
Calculates the difference in R^2 that results from omitting a predictor for each predictor as described by Fischer(2015).
}
\examples{
data=data.frame(y=rnorm(100)+scale(1:100),x1=rnorm(100)+scale(1:100),x2=rnorm(100));
l=lm(y~.,data=data)
deltas(l,data[c("x1","x2")],data["y"])
}
\references{
Fischer (2015). How to determine the unique contributions of input-variables to the nonlinear regression function of a multilayer perceptron. Ecological Modelling, 309, 60-63.
}
\keyword{modeling}
