% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptionsLDA.r
\name{descriptionsLDA}
\alias{descriptionsLDA}
\title{Function descriptionsLDA}
\usage{
descriptionsLDA(lda, n = 5)
}
\arguments{
\item{n}{Numeric value specifying the number of terms to describe each topic.}

\item{lds}{topicmodels::LDA}
}
\description{
Returns topic descriptions.
}
\details{
Returns topic descriptions.
}
\keyword{plotting}
