% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.r
\name{dotplot}
\alias{dotplot}
\title{Function dotplot}
\usage{
dotplot(x = NULL, relative = T, vertLine = T, horizLines = T,
  max = 40, main = NULL, pch = 16, cex = 0.7, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{relative}{Logical value. Defaults to T.}

\item{vertLine}{Logical value specifying whether to plot a vertical line at x=0. Defaults to T.}

\item{horizLines}{Logical value specifying whether to plot horizontal lines between each point and the y axis. Defaults to T.}

\item{max}{Numeric value indicating the maximum number of frequencies to be displayed. Defaults to 40.}

\item{main}{Character value specifying the plot's title. Defaults to "Comparison of Word Frequencies"}

\item{pch}{Numeric element speccifying the symbol to represent points. Defaults to 16}

\item{cex}{Numeric value indicating the text size. Defaults to .7.}
}
\description{
Plots a modified dotchart.
}
\details{
Plots vector as modified dotchart.
}
\examples{
dotplot(rnorm(100))
}
\keyword{plotting}
