% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusivityLDA.r
\name{exclusivityLDA}
\alias{exclusivityLDA}
\title{Function exclusivityLDA}
\usage{
exclusivityLDA(lda, dtm, lambda = 0, num.words = 0)
}
\arguments{
\item{lda}{Object of class LDA.}

\item{DTM}{document -term matrix.}
}
\description{
Calculates exclusivity of an LDA.
}
\examples{

}
\keyword{modeling}
