% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTable.r
\name{extractTable}
\alias{extractTable}
\title{Function extractTable}
\usage{
extractTable(x, reg.up = NULL, reg.down = NULL, reg.left = "^",
  reg.right = "$", reg.fix = NULL, convert = F,
  correctNotation = T)
}
\arguments{
\item{x}{A character vector containing one text-line per element.}
}
\description{
Extracts table as data.frames from multiple lines of text.
}
\details{
Extracts table as data.frames from multiple lines of text. Columns are assumed to be separated by whitespaces that are placed at the same position in each line. In its current state this function has to be considered experimental. Returns a numeric data.frame. It is highly recommended to check the attribute "raw" that's attached to the data.frame.
}
\keyword{scraping}
