% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.r
\name{first}
\alias{first}
\title{Function first}
\usage{
first(x, s = 0, factorsAsStrings = T)
}
\arguments{
\item{x}{A vector or list.}
}
\description{
Returns the first element of a vector.
}
\details{
Returns the first element of a vector.
}
\examples{
first(c(1,2,3))
}
\keyword{preprocessing}
