% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowerplot.r
\name{flowerplot}
\alias{flowerplot}
\title{Function flowerplot}
\usage{
flowerplot(x = NULL, maximum = NULL, rownames = NULL,
  colnames = NULL, main = NULL, color = NULL, color2 = "lightgrey",
  add.numbers = F, ncex = 0.8, ndigits = 1, ncol = "black",
  cex = 0.8, cex.legend = 0.8, xlim = NULL, ylim = NULL,
  dist = 4, legend = NULL, xshift = 0)
}
\arguments{
\item{x}{Numeric data.frame or matrix containing the values to be displayed.}

\item{maximum}{Numeric value representing the maximum value. Only needs to be specified when the data does not contain the theoretically possible maximum.}

\item{rownames}{Character vector of the same length as x, containing the row names to be displayed. If NULL (default) rownames of x are applied.}

\item{colnames}{Character vector of the same length as x, containing the column names to be displayed. If NULL (default) colnames of x are applied.}

\item{main}{Character value, containing the title to be displayed. Defaults to NULL.}

\item{color}{Character vector, containing the colors of petals. If NULL (default) the rainbow palette is applied.}

\item{color2}{Character value, containing the color of the background petals. If NULL, no background petals are plotted. Defaults to "lightgrey".}

\item{add.numbers}{Logical value specifying whether to draw numbers next to each petal. Defaults to F.}

\item{ncex}{Size of fonts. If NA (default) is set to cex.}

\item{ndigits}{Numeric value specifying the number of digits to be plotted (if add.numbers==T). Defaults to 2.}

\item{ncol}{Vector containing the color of bars. Defaults to "black".}

\item{xshift}{Numeric value specifying how much to shift flowers to the right. Defaults to 0.}
}
\description{
Plots numeric matrix as a field of flowers.
}
\details{
Plots data.frame as a field of flowers. Each column is represented as a separate flower, each row as a flower's petal.
}
\examples{
flowerplot()
}
\keyword{plotting}
