% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFile.r
\name{getFile}
\alias{getFile}
\title{Function getFile}
\usage{

  getFile(url = "http://www.bamf.de/SharedDocs/Anlagen/DE/Publikationen/Flyer/flyer-schluesselzahlen-asyl-halbjahr-2018.pdf?__blob=publicationFile",
  filename = NULL, mode = "wb", format = "pdf")
}
\arguments{
\item{url}{A valid URL.}

\item{filename}{Character value specifying the path to save the downloaded file to.}
}
\description{
Downloads File from URL.
}
\details{
Downdloads File from URL.
}
\examples{
getFile()
}
\keyword{scraping}
