% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHTML.r
\name{getHTML}
\alias{getHTML}
\title{Function getHTML}
\usage{

  getHTML(url = "https://scholar.google.de/citations?user=-TjY7oEAAAAJ&hl=de&oi=sra",
  encoding = "UTF-8", save = F, prefix = Sys.Date(), closeCon = T,
  method = "libcurl", silent = T)
}
\arguments{
\item{url}{A valid URL.}

\item{encoding}{Character value specifying the encoding. Defaults to "UTF-8".}

\item{save}{Logical value specifying whether the HTML-code should be saved to a csv-file.}

\item{prefix}{Character value specifying the beginning of the filename (in case of saving). Defaults to Sys.Date().}
}
\description{
Extracts HTML from an URL.
}
\details{
Extracts HTML from an URL. Assumes UTF-8 encoding by default. Returns a character element containing HTML-code.
}
\examples{
getHTML()
}
\keyword{scraping}
