% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gramcloud.r
\name{gramcloud}
\alias{gramcloud}
\title{Function gramcloud}
\usage{
gramcloud(x = NULL, ngram = 3, min.freq = 1, equal.shares = T,
  delete = T)
}
\arguments{
\item{x}{Character vector containing the text of documents.}
}
\description{
Plots wordcloud based on words and/or ngrams.
}
\details{
Plots wordcloud based on words and/or ngrams. By default words that are part of plotted ngrams are not plotted (delete=T).
}
\examples{
gramcloud(c("This is an exemplary text","This is too","And this is probably the longest text"),ngram=2,min.freq=1)
}
\keyword{(F).}
\keyword{(T)}
\keyword{1.}
\keyword{2.}
\keyword{Defaults}
\keyword{Logical}
\keyword{Numeric}
\keyword{T.}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{are}
\keyword{be}
\keyword{constituting}
\keyword{corresponding}
\keyword{delete}
\keyword{deleted}
\keyword{determines}
\keyword{equal.shares}
\keyword{for}
\keyword{frequent}
\keyword{higher}
\keyword{if}
\keyword{indicating}
\keyword{is}
\keyword{least}
\keyword{lower}
\keyword{maximum}
\keyword{min.freq}
\keyword{minimum}
\keyword{ngram}
\keyword{ngram-order}
\keyword{ngrams}
\keyword{number}
\keyword{occurences}
\keyword{of}
\keyword{or}
\keyword{order}
\keyword{part}
\keyword{plotted.}
\keyword{plotting}
\keyword{relative}
\keyword{same}
\keyword{should}
\keyword{size}
\keyword{that}
\keyword{the}
\keyword{they}
\keyword{to}
\keyword{tokens}
\keyword{value}
\keyword{whether}
\keyword{word}
