% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kCTM.r
\name{kCTM}
\alias{kCTM}
\title{Function kCTM}
\usage{
kCTM(data = NULL, d = NULL, K = c(2:10), method = "VEM",
  removestopwords = F, lower.thresh = 0, seed1 = 0,
  control = list(nstart = 1, best = T, burnin = 4000, iter = 500, thin =
  500, seed = NULL), plot = T, ...)
}
\description{
Function kCTM
}
\keyword{modeling}
