% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLDA.r
\name{kLDA}
\alias{kLDA}
\title{Function kLDA}
\usage{
kLDA(data = NULL, K = c(2:10), dtm = NULL, method = "Gibbs",
  removestopwords = F, lower.thresh = 0, seed1 = 0,
  control = list(nstart = 1, best = T, burnin = 4000, iter = 500, thin =
  500, seed = NULL), plot = T, ...)
}
\description{
Function kLDA
}
\keyword{modeling}
