% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSTM.r
\name{kSTM}
\alias{kSTM}
\title{Function kSTM}
\usage{
kSTM(data, K = c(2:10), init.type = "LDA", removestopwords = F,
  lower.thresh = 0, seed1 = 0, plot = T, ...)
}
\description{
Function kSTM
}
\keyword{modeling}
