% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelbreaker.r
\name{labelbreaker}
\alias{labelbreaker}
\title{Function textbreaker}
\usage{
labelbreaker(x, length = 30, separator = "\\n     ", trim = F, ...)
}
\description{
Adds linebreaks to each element of a character vector.
}
\details{
Adds linebreaks to each element of a character vector.
}
\examples{
textbreaker("hello world hello world hello world hello world")
}
\keyword{helper}
