% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.r
\name{last}
\alias{last}
\title{Function last}
\usage{
last(x, s = 0, factorsAsStrings = T)
}
\arguments{
\item{x}{A vector or list.}
}
\description{
Returns the last element of a vector.
}
\details{
Returns the last element of a vector.
}
\examples{
last(c(1,2,3))
}
\keyword{preprocessing}
