% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCorpus.r
\name{loadCorpus}
\alias{loadCorpus}
\title{Function loadCorpus}
\usage{
loadCorpus(path = "2017-12-29 corpus.txt")
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Function loadCorpus
}
\keyword{mining}
\keyword{text}
