% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchAll.r
\name{matchAll}
\alias{matchAll}
\title{Function matchAll}
\usage{
matchAll(string = NA, pattern = ".*", value = T, ...)
}
\arguments{
\item{string}{Character vector where matches are sought.}

\item{pattern}{Character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector.}

\item{value}{Logical value. If T (default) returns substrings. Otherwise returns positions.}
}
\description{
Search for matches to argument pattern within each element of a character vector. Returns corresponding substrings (compare stringr::str_match_all).
}
\examples{
matchAll(string=c("Hello world","Hello"),pattern="wo.*")
}
\keyword{helper}
