% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchOne.r
\name{matchOne}
\alias{matchOne}
\title{Function matchOne}
\usage{
matchOne(string = NA, pattern = ".*", value = T, ...)
}
\arguments{
\item{string}{Character vector where matches are sought.}

\item{pattern}{Character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector.}

\item{value}{Logical value. If T (default) returns substrings. Otherwise returns positions.}
}
\description{
Search for first match to argument pattern within each element of a character vector. Returns corresponding substrings (compare stringr::str_match).
}
\examples{
matchOne(string=c("Hello world","Hello"),pattern="wo.*")
}
\keyword{helper}
