% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.analysis.r
\name{outlier.analysis}
\alias{outlier.analysis}
\title{Function outlier.anaysis}
\usage{
outlier.analysis(df = NULL, cutoff = NULL, plot = T,
  col = rgb(106/255, 160/255, 186/255), varwidth = T, outline = T,
  main = "Outlier Analysis", cex = 0.8, ...)
}
\arguments{
\item{df}{A data.frame containing numeric vectors.}
}
\description{
Determines outliers based on boxplots mahalanobis distance.
}
\details{
Determines outliers based on boxplots and mahalanobis distance. Returns boxplots with additional attribute "mahalanobis"
}
\examples{
outlier.analysis(data.frame(c(1,2,100)))
}
\keyword{preprocessing}
