% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDF.r
\name{plotDF}
\alias{plotDF}
\title{Function plotDF}
\usage{
plotDF(data = NULL, col = NULL, ...)
}
\arguments{
\item{data}{Numeric data.frame.}

\item{col}{Character vector specifying each plot's color. If NULL (default) the rainbow palette is applied.}
}
\description{
Plots each vector of a data.frame.
}
\details{
Plots each vector of a data.frame either as histogram or as barplot.
}
\examples{
plotDF()
}
\keyword{plotting}
