% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFA.r
\name{plotFA}
\alias{plotFA}
\title{Function plotFA}
\usage{
plotFA(fa = NULL)
}
\arguments{
\item{data}{Numeric data.frame.}
}
\description{
Plots psych::fa object.
}
\details{
Plots psych::fa object.
}
\examples{
af.fa(data.frame(x=rnorm(100),y=rnorm(100)+scale(1:100),z=rnorm(100)+scale(1:100)));plotFA(fa)
}
\keyword{plotting}
