% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLDA.r
\name{plotLDA}
\alias{plotLDA}
\title{Function plotLDA}
\usage{
plotLDA(lda = NULL, labels = NULL,
  main = "Topic distribution over documents", border = NA, space = 0,
  stopwords = NULL, cex = 0.7)
}
\arguments{
\item{lda}{Object of class LDA.}

\item{labels}{character vector with document labels.}
}
\description{
Plots topic document distribution and topic descriptions.
}
\details{
Plots topic document distribution and topic descriptions based on a topicmodels::LDA object.
}
\examples{
l=af.lda(c("Hello world","Hallo Welt"));plotLDA(l)
}
\keyword{plotting}
