% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLM.r
\name{plotLM}
\alias{plotLM}
\title{Function plotLM}
\usage{
plotLM(lm = NULL, ...)
}
\arguments{
\item{lm}{Object of class lm.}
}
\description{
Plots regression coefficiens based on lm object.
}
\details{
Plots regression coefficiens based on lm object.
}
\examples{
plotLM()
}
\keyword{plotting}
