% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMAT.r
\name{plotMAT}
\alias{plotMAT}
\title{Function plotMAT}
\usage{
plotMAT(matrix = NULL, main = "Cumulation over Time", xlab = "",
  ylab = "cumulated sum", xlim = NULL, ylim = NULL,
  xlim.factor = 1.5, las = 1, srt = 45, cumsum = T,
  show.legend = F, add.shadow = F, add.grid = T, col = NULL,
  cex = 0.7, frame = 1, manual.addon = NULL)
}
\arguments{
\item{matrix}{Numeric matrix containing the values to be displayed.}

\item{main}{Character value representing the barplot's title. Defaults to "Cumulation over time"}

\item{xlab}{Character value representing the x-axis label. Defaults to "time"}

\item{xlim}{Numeric vector with two elements. If NULL (default) xlim is detemined automatically.}

\item{ylim}{Numeric vector with two elements. If NULL (default) ylim is detemined automatically.}

\item{xlim.factor}{Numeric value for adding extra space to the right of the plot. Defaults to 1.}

\item{las}{Numeric value specifying the rotation of the y-axis (0 for 90%-rotation, 1 for 0% rotation). Defaults to 1.}

\item{srt}{Numeric value specifying the rotation of the x-axis (between 0 and 360 degrees). Defaults to 45.}

\item{cumsum}{Logical value indicating whether the cumsum of each row in the matrix of each row should be plotted. Defaults to T.}

\item{show.legend}{Logical value indicating whether a legend should be drawn instead of texts. Defaults to T.}

\item{add.shadow}{Logical value indicating whether lines should be surrounded ba a black line. Defaults to F.}

\item{add.grid}{Logical value indicating whether a grid should be drawn. Defaults to F.}

\item{col}{Vector containing each line's color. If NULL (default) colors are generated based on the rainbow-palette.}

\item{cex}{Relative size of fonts. Defaults to .7.}

\item{frame}{Relative size of invisible frame around fonts. Defaults to 1.}

\item{manual.addon}{Numeric vector containing verical adjustments.}

\item{main}{Character value representing the y-axis label. Defaults to "cumulated sum"}
}
\description{
Plots a numeric matrix as one line per row.
}
\details{
Plots a numeric matrix as one line per row. By default cumsum of each row is plotted.
}
\examples{
plotMAT()
}
\keyword{plotting}
