% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNNET.r
\name{plotNNET}
\alias{plotNNET}
\title{Function plotNNET}
\usage{
plotNNET(net = NULL)
}
\arguments{
\item{net}{Object of class nnet.}
}
\description{
Plots the structure of a nnet object.
}
\details{
Plots the structure of a nnet object.
}
\examples{
library(nnet);net=nnet(y~.,data=data.frame(y=rnorm(100),x=rnorm(100)),size=3,trace=F);plotNNET(net)
}
\keyword{plotting}
