% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSTM.r
\name{plotSTM}
\alias{plotSTM}
\title{Function plotSTM}
\usage{
plotSTM(stm = NULL, labels = NULL,
  main = "Topic distribution over documents", border = NA, space = 0,
  stopwords = NULL, cex = 0.7, srt = 45, plot = T)
}
\arguments{
\item{stm}{Object of class stm.}

\item{labels}{character vector with document labels.}

\item{srt}{Numeric value specifying the rotation of the x-axis (between 0 and 360 degrees). Defaults to 45.}
}
\description{
Plots topic document distribution and topic descriptions based on an stm object.
}
\details{
Plots topic document distribution and topic descriptions based on an stm object.
}
\examples{
s=af.stm(c("Hello world","Hallo Welt"));plotSTM(s)
}
\keyword{plotting}
