% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXY.r
\name{plotXY}
\alias{plotXY}
\title{Function plotXY}
\usage{
plotXY(x = NULL, y = NULL, complexity = 1, na.rm = T,
  color1 = rgb(0, 0, 0, 0.7), color2 = rgb(0, 0, 1), color3 = rgb(0,
  0, 1, 0.2), xlab = "x", ylab = "y", main = "Bivariate Relation",
  sub = "Shaded area represents 95\%-prediction interval.", pch = 16,
  lwd = 2, cex = 0.7, cex.sub = 0.7, generalize = F, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector of the same length as x.}

\item{na.rm}{Logical value indicating whether missing values should be skipped. Defaults to T.}

\item{color1}{Color of points in the scattergram. Defaults to rgb(0,0,0,.7).}

\item{color2}{Color of the regression line. Defaults to rgb(0,0,1).}

\item{color3}{Color of the prediction interval. Defaults to rgb(0,0,1,.2).}
}
\description{
Plots bivariate correlation based on two numeric vectors.
}
\details{
Plots scattergram and bivariate correlation based on two numeric vectors.
}
\examples{
plotXY()
}
\keyword{plotting}
