% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorSTM.r
\name{posteriorSTM}
\alias{posteriorSTM}
\title{Function posteriorSTM}
\usage{
posteriorSTM(tm)
}
\arguments{
\item{tm}{topic model object of class stm or LDA.}
}
\description{
Posterior for STM topic models.
}
\examples{

}
\keyword{modeling}
