% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCodeFrequencies.r
\name{predictCodeFrequencies}
\alias{predictCodeFrequencies}
\title{Function predictCodeFrequencies}
\usage{
predictCodeFrequencies(corpus, predictor, binary = T, min = NULL,
  mostFrequent = 10, plot = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Correlates code frequencies across documents with a numeric vector.
}
\details{
Correlates code frequencies across documents with a numeric vector.
}
\examples{
correlateWordFrequencies(c("<1>hello</1> world"),c("<1>hello</1> world","<2>hallo</2>"))
}
\keyword{mining}
\keyword{text}
