% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictWordFrequencies.r
\name{predictWordFrequencies}
\alias{predictWordFrequencies}
\title{Function predictCodeFrequencies}
\usage{
predictWordFrequencies(
  corpus,
  predictor,
  binary = T,
  min = NULL,
  mostFrequent = 10,
  plot = T
)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Correlates word frequencies across documents with a numeric vector.
}
\details{
Correlates word frequencies across documents with a numeric vector.
}
\examples{
correlateWordFrequencies(c("<1>hello</1> world"),c("<1>hello</1> world","<2>hallo</2>"))
}
\keyword{mining}
\keyword{text}
