% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDF.r
\name{prepareDF}
\alias{prepareDF}
\title{Function prepareDF}
\usage{
prepareDF(data, output = NULL, scale = T, out.omit = T,
  na.omit = T)
}
\arguments{
\item{data}{data.frame containing data.}

\item{output}{index or name of criterion in data.}

\item{out.omit}{logical value, specifying whether outliers (scale(data)>3) shoud be set to NA.}

\item{na.omit}{logical value, specifying whether NA shoud be omitted.}
}
\description{
Scales numeric data and dummy codes factors, omits outliers and missing values.
}
\examples{
data=data.frame(y=rnorm(100),x1=rnorm(100),x2=rnorm(100));
prepareDF(data,"y1")
}
\keyword{preprocessing}
