% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityLDA.r
\name{qualityLDA}
\alias{qualityLDA}
\title{Function qualityLDA}
\usage{
qualityLDA(lda, dtm, plot = F, n = 5, type = 1)
}
\arguments{
\item{lda}{Object of class LDA (or list of LDA objects).}

\item{dtm}{document-term matrix.}

\item{n}{Parameter to be passed to accuracyLDA. Number of top-terms to search for in the documents. Defaults to 5.}

\item{type}{Parameter to be passed to accuracyLDA. Determines how the results are aggregated. Defaults to 1. If type==0 the function returns the number of each document's top-topic's n top-terms that are actually present in the document as a vector e; type==1 returns mean(e>0); type==2 returns mean(e==n); type==3 returns mean(e);  type==4 returns median(e>0); type==5 returns median(e==n); type==6 returns median(e);  type==7 returns min(e>0); type==8 returns min(e==n); type==9 returns min(e);}
}
\description{
Calculates and plots accuracyLDA, logLik, AIC and BIC for one or more LDA-objects.
}
\examples{
accuracyLDA(lda,dtm)
}
\keyword{modeling}
