% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDF.r
\name{reduceDF}
\alias{reduceDF}
\title{Function reduceDF}
\usage{
reduceDF(data, output = NULL)
}
\arguments{
\item{data}{data.frame containing data.}

\item{output}{index or name of criterion in data.}

\item{y}{numeric vector.}
}
\description{
Applies LASSO regression on predictors and their criterion.
}
\examples{
data=data.frame(y=rnorm(100),x1=rnorm(100),x2=rnorm(100));
reduceDF(data,"y1")
}
\keyword{preprocessing}
