% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveAllCodes.r
\name{retrieveAllCodes}
\alias{retrieveAllCodes}
\title{Function retrieveAllCodes}
\usage{
retrieveAllCodes(corpus)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Retrieves all codes in a character vector.
}
\details{
Retrieves all codes in a character vector.
}
\examples{
retrieveAllCodings("<1>hello</1> world")
}
\keyword{mining}
\keyword{text}
