% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveAllCodings.r
\name{retrieveAllCodings}
\alias{retrieveAllCodings}
\title{Function retrieveAllCodings}
\usage{
retrieveAllCodings(corpus, context = F)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Retrieves all codings in a character vector.
}
\details{
Retrieve all codings in a character vector.
}
\examples{
retrieveAllCodings("<1>hello</1> world")
}
\keyword{mining}
\keyword{text}
