% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveCodeFrequencies.r
\name{retrieveCodeFrequencies}
\alias{retrieveCodeFrequencies}
\title{Function retrieveCodeFrequencies}
\usage{
retrieveCodeFrequencies(corpus, byString = T, binary = F, plot = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Retrieves all codes in a character vector and returns frequencies of codings.
}
\details{
Retrieves all codes in a character vector and returns frequencies of codings.
}
\examples{
retrieveCodeFrequencies("<1>hello</1> world")
}
\keyword{mining}
\keyword{text}
