% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveCoding.r
\name{retrieveCoding}
\alias{retrieveCoding}
\title{Function retrieveCoding}
\usage{
retrieveCoding(corpus, coding)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Retrieves certain codings.
}
\details{
Retrieves certain codings.
}
\examples{
retrieveCoding("<h1>hello</h1> world","h1")
}
\keyword{mining}
\keyword{text}
