% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveCodingPosition.r
\name{retrieveCodingPosition}
\alias{retrieveCodingPosition}
\title{Function retrieveCodingPosition}
\usage{
retrieveCodingPosition(corpus, coding)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Function retrieveCodingPosition
}
\keyword{mining}
\keyword{text}
