% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrievePassage.r
\name{retrievePassage}
\alias{retrievePassage}
\title{Function retrievePassage}
\usage{
retrievePassage(corpus, passage, context = F)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Function retrievePassage
}
\keyword{mining}
\keyword{text}
