% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrievePassagePosition.r
\name{retrievePassagePosition}
\alias{retrievePassagePosition}
\title{Function retrievePassagePosition}
\usage{
retrievePassagePosition(corpus, passage, context = F)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Function retrievePassagePosition
}
\keyword{mining}
\keyword{text}
