% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCorpus.r
\name{saveCorpus}
\alias{saveCorpus}
\title{Function saveCorpus}
\usage{
saveCorpus(corpus)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Function saveCorpus
}
\keyword{mining}
\keyword{text}
