% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDevs.r
\name{saveDevs}
\alias{saveDevs}
\title{Function saveDevs}
\usage{
saveDevs(dev = png, prefix = NA, width = NA, height = NA,
  units = "in", res = 300, ...)
}
\arguments{
\item{dev}{Graphical device (currently, pdf, png, bmp, jpg and tiff are supported).}

\item{width}{width of the devices. If NA (default), width of the devices is left unchanged.}

\item{units}{Parameter passed to dev.copy. Defaults to "in". Is ignored if dev==pdf.}

\item{res}{Resolution (dots per inch) of the devices. Defaults to 300. Is ignored if dev==pdf.}

\item{width}{height of the devices. If NA (default), height of the devices is left unchanged.}
}
\description{
Saves each element of dev.list() to a graphical device such as pdf, png or similar.
}
\details{
Saves each element of dev.list() to a graphical device such as pdf, png or similar.
}
\examples{
saveDevs();
}
\keyword{plotting}
