% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeHTML.r
\name{scrapeHTML}
\alias{scrapeHTML}
\title{Function scrapeHTML}
\usage{
scrapeHTML(html, short = F, edit = T, save = F, plot = F,
  filename = NULL)
}
\arguments{
\item{html}{A character element containing HTML-code.}

\item{short}{Logical value specifying whether only lines with verbal information or link should be returned. Defaults to F.}

\item{edit}{Logical value specifying whether the data.frame should be plotted/edited.}

\item{save}{Logical value specifying whether the HTML-code should be saved to a csv-file.}

\item{plot}{Logical value specifying whether to plot the frequency of each HTML-tag found in the html-object.}

\item{filename}{Character value specifying the filename (if save is TRUE). If NULL (default) as.numeric(Sys.time()) is applied.}
}
\description{
Extracts Information from HTML.
}
\details{
Extracts Information from HTML code (as returned by quantqual::getHTML, for example). Returns a data.frame with three columns: the first column contains html-code, the second column contains extracted verbal information, and the third column contains extracted links.
}
\examples{
scrapeHTML(getHTML())
}
\keyword{scraping}
