% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeJDDM.r
\name{scrapeJDDM}
\alias{scrapeJDDM}
\title{Function scrapJDDM}
\usage{
scrapeJDDM(wd = NULL, saveHTML = F)
}
\arguments{
\item{wd}{Pathname for the working directory to save the files into.}
}
\description{
Downloads data and metadata from the database of the Journal of Dynamic Decision Making to the working directory.
}
\details{
Downloads data and metadata from the database of the Journal of Dynamic Decision Making to the working directory. Returns a data.frame containing the relevant information.
}
\examples{
s=scrapeJDDM();
}
\keyword{JDDM}
