% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetHTML.r
\name{subsetHTML}
\alias{subsetHTML}
\title{Function subsetHTML}
\usage{
subsetHTML(html, pattern = "class=\\"m-resultListEntries__content\\"",
  start = "<div", end = "</div>", save = T, prefix = Sys.Date())
}
\arguments{
\item{html}{A character element containing HTML-code.}

\item{pattern}{Regular expression specifying the tag to be used for subsetting.}

\item{start}{Regular expression. Defaults to "<div".}

\item{end}{Regular expression. Defaults to "</div>".}

\item{save}{Logical value specifying whether the HTML-code should be saved to a csv-file.}

\item{prefix}{Character value specifying the beginning of the filename (in case of saving). Defaults to Sys.Date().}
}
\description{
Extracts a coherent subset of code from HTML-code.
}
\details{
Extracts a coherent subset of code from HTML-code.
}
\examples{
subsetHTML(getHTML("https://jobs.meinestadt.de/nuernberg/suche?words=Wissenschaftlicher\%20Mitarbeiter"))
}
\keyword{scraping}
