% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCodeFrequencies.r
\name{summarizeCodeFrequencies}
\alias{summarizeCodeFrequencies}
\title{Function summarizeCodeFrequencies}
\usage{
summarizeCodeFrequencies(corpus, plot = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Summarizes code frequencies based on a character vector.
}
\details{
Summarizes code frequencies based on a character vector.
}
\examples{
summarizeCodeFrequencies("<1>hello</1> world")
}
\keyword{mining}
\keyword{text}
