% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeWordFrequencies.r
\name{summarizeWordFrequencies}
\alias{summarizeWordFrequencies}
\title{Function summarizeWordFrequencies}
\usage{
summarizeWordFrequencies(corpus, plot = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Summarizes word frequencies based on a character vector.
}
\details{
Summarizes word frequencies based on a character vector.
}
\examples{
summarizeWordFrequencies("<1>hello</1> world")
}
\keyword{mining}
\keyword{text}
