% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textbreaker.r
\name{textbreaker}
\alias{textbreaker}
\title{Function textbreaker}
\usage{

  textbreaker(text = "Let's write some text with many different words and plot it in a nice little device.",
  zeilenabstand = 1, maxlength = 30, size = 1, centered = F,
  separator = "\\n     ", plot = F)
}
\description{
Adds linebreaks to text.
}
\details{
Adds linebreaks to text.
}
\examples{
textbreaker("hello world hello world hello world hello world")
}
\keyword{helper}
