% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfIdf.r
\name{tfIdf}
\alias{tfIdf}
\title{Function tfIdf}
\usage{
tfIdf(dtm, doc)
}
\arguments{
\item{dtm}{Document Term Matrix.}

\item{doc}{Numeric element specifying a document in the dtm.}
}
\description{
Term-Frequency-Inverse-Term-Frequency.
}
\keyword{mining}
\keyword{text}
