% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.r
\name{trim}
\alias{trim}
\title{Function trim}
\usage{
trim(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Trims elements of a character vector.
}
\details{
Trims elements of a character vector.
}
\examples{
trim(" Hi  ")
}
\keyword{preprocessing}
