% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncodeCode.r
\name{uncodeCode}
\alias{uncodeCode}
\title{Function uncodeCode}
\usage{
uncodeCode(corpus, code = "c1", mess = T, assign = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Removes all codings of a certain code from a character vector.
}
\details{
Removes all codings of a certain code from a character vector.
}
\examples{
uncodeCode("<1>hello</1> world")
}
\keyword{mining}
\keyword{text}
