% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncodeCodingNo.r
\name{uncodeCodingNo}
\alias{uncodeCodingNo}
\title{Function uncodeCorpus}
\usage{
uncodeCodingNo(corpus, codingNo = 1, mess = T, assign = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Function uncodeCorpus
}
\keyword{mining}
\keyword{text}
