% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncodeCorpus.r
\name{uncodeCorpus}
\alias{uncodeCorpus}
\title{Function uncodeCorpus}
\usage{
uncodeCorpus(corpus, mess = T, assign = T, tolower = F)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Removes all codings from a character vector.
}
\details{
Removes all codings from a character vector.
}
\examples{
uncodeCorpus("<1>hello</1> world")
}
\keyword{mining}
\keyword{text}
