% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncodePassage.r
\name{uncodePassage}
\alias{uncodePassage}
\title{Function uncodePassage}
\usage{
uncodePassage(corpus, passage, mess = T, assign = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Function uncodePassage
}
\keyword{mining}
\keyword{text}
