% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncodePosition.r
\name{uncodePosition}
\alias{uncodePosition}
\title{Function uncodeCorpus}
\usage{
uncodePosition(corpus, position, mess = T, assign = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Function uncodeCorpus
}
\keyword{mining}
\keyword{text}
