% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undoTwoColumns.r
\name{undoTwoColumns}
\alias{undoTwoColumns}
\title{Function undoTwoColumns}
\usage{
undoTwoColumns(doc, sepwidth = 2, mid = NULL, regex = NULL,
  trim = T)
}
\arguments{
\item{doc}{Character element with text with two columns (separated by whitespace).}

\item{sepwidth}{Numeric element specifying the number of whitespaces that are assumed to separate columns. Defaults to 2.}

\item{mid}{Numeric element specifying the number of characters until the end of the first column. If NULL (default) this number is estimated based on the text.}

\item{trim}{Logical element speficying whether the text is trimmed before inferring that an empty line is to be treated as/like a pagebreak.}
}
\description{
Returns one-column version of two column-text.
}
\details{
Returns one-column version of two column-text.
}
\keyword{scraping}
