% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecToDTM.r
\name{vecToDTM}
\alias{vecToDTM}
\title{Function vecToDTM}
\usage{
vecToDTM(corpus = "hello,  \\nworld", stopwords = NULL,
  lowerCase = T, min = 0, minDocs = 0, sort = F, plot = F,
  remove_punct = T, ...)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Document-Term matrix based on character vector.
}
\details{
Document-Term matrix based on character vector.
}
\examples{
vecToDTM("hello world")
}
\keyword{mining}
\keyword{text}
