% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordCountCorpus.r
\name{wordCountCorpus}
\alias{wordCountCorpus}
\title{Function wordCountCorpus}
\usage{
wordCountCorpus(corpus, lowerCase = T, min = 0, sort = T, plot = T)
}
\arguments{
\item{corpus}{Character vector containing one document per element.}
}
\description{
Returns word frequencies based on a character vector.
}
\details{
Returns word frequencies based on a character vector.
}
\examples{
wordCountCorpus("<1>hello</1> world")
}
\keyword{mining}
\keyword{text}
