% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeReport.r
\name{writeReport}
\alias{writeReport}
\title{Function writeReport}
\usage{
writeReport(code = list(function() {     library(quantqual)     plotMAT()
      bp()     flowerplot()     spiderplot() }, function() {    
  "Hello world" }),
  intro = c("# Report\\nThe research report was generated using the \\"quantqual\\" package for GNU R."),
  options = "fig.height=7, fig.width=7", author = NULL,
  title = "Research Report", format = "word_document",
  date = "`r Sys.Date()`")
}
\arguments{
\item{code}{List of functions containing the code to be documented in the document.}

\item{intro}{Character vector containing the text to be plotted above each chunk of code.}

\item{options}{Character vector containing the options to be applied for each chunk of code.}

\item{author}{Character element specifying the author of the document. Defaults to "`r Sys.Date()`".}

\item{title}{Character element specifying the document title.}

\item{format}{Character element specifying the output format (e.g., "word_document", "html_document" or  "pdf_document").}

\item{date}{Character element specifying the date of the document. Defaults to "Mr. X".}
}
\description{
Generates and renders rmarkdown-documents.
}
\details{
Generates and renders rmarkdown-documents.
}
\examples{
writeReport(data.frame(group1=c(variable1=1,variable2=2),group2=c(variable1=3,variable2=4)),horiz=T,main="Example")
}
\keyword{plotting}
