% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.lda.r
\name{af.lda}
\alias{af.lda}
\title{Function af.lda}
\usage{
af.lda(data = c("Hallo Welt", "Hello World"), k = 3,
  method = "Gibbs", nstart = 10, seed = 0)
}
\description{
Function af.lda
}
\keyword{modeling}
