% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxedText.r
\name{boxedText}
\alias{boxedText}
\title{Function boxedText}
\usage{
boxedText(
  x,
  y,
  text = "Hello \\nWorld",
  x2 = NULL,
  y2 = NULL,
  maxlength = NULL,
  col = "yellow",
  border = "black",
  pos = NULL,
  font = 1,
  hspace = 1,
  vspace = 1.2,
  cex = 1,
  decollide = F,
  decollide.repetitions = 10,
  ...
)
}
\arguments{
\item{x}{Numeric value specifying the text's position on the x-axis.}

\item{y}{Numeric value specifying the text's position on the y-axis.}

\item{x2}{Numeric value specifying the arrow's position on the x-axis.}

\item{y2}{Numeric value specifying the arrow's position on the y-axis.}

\item{col}{Character value specifying the color of the border. Defaults to "black".}

\item{pos}{Numeric value specifying the position of the arrow relative to the box (0=no arrow plotted, 1=down, 2=left, 3=up, 4=right).}

\item{...}{Additional graphical parameters for text.}

\item{txt}{Character value specifying the text to be plotted.}
}
\description{
Adds a boxed text to a plot.
}
\details{
Adds a boxed text to a plot.
}
\examples{
plot(1:10,1:10);boxedText(7,2.5,text="This (3;3) is a remarkable point worth emphasizing",3,3,col="yellow",maxlength=20)
}
\keyword{plotting}
