% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.r
\name{bp}
\alias{bp}
\title{Function bp}
\usage{
bp(x = NULL, sd = NULL, cex = 0.7, beside = T, horiz = F,
  add.numbers = F, ndigits = 2, grid = T, plot = T, main = NULL,
  xlim.factor = 1.5, las = 1, srt = 45, xlim = NULL,
  names.arg = NULL, legend.text = NULL, col = NULL, axes = T,
  add = F, ...)
}
\arguments{
\item{x}{Numeric vector, matrix or data.frame containing the values to be displayed.}

\item{sd}{Numeric vector, matrix or data.frame of same format as x containing standard deviations.}

\item{cex}{Size of fonts. Defaults to .7.}

\item{beside}{Logical value indicating if bars should be placed next to each other. Defaults to T.}

\item{horiz}{Logical value indicating if bars should be placed horizontal. Defaults to F.}

\item{add.numbers}{Logical value indicating if numbers should be placed above bars. Defaults to F.}

\item{ndigits}{Numeric value specifying the number of digits to be plotted (if add.numbers==T). Defaults to 2.}

\item{grid}{Logical value indicating whether to plot a grid. Defaults to T.}

\item{plot}{Logical value indicating whether to plot the barplot. Defaults to T.}

\item{main}{Character vector with one element containing the barplot's title. Defaults to NULL}

\item{xlim.factor}{Numeric value for adding extra space to the right of the barplot. Defaults to 1.}

\item{las}{Numeric value specifying the rotation of the y-axis (0 for 90 percent rotation, 1 for 0 percent rotation). Defaults to 1.}

\item{srt}{Numeric value specifying the rotation of the x-axis (between 0 and 360 degrees). Defaults to 45.}

\item{xlim}{Numeric vector containing xlim.}

\item{names.arg}{Character vector containing names of bars. If NULL (default) colnames of x will be applied as names.arg.}

\item{legend.text}{Legend text. Set to NA or to F to supress legend. If NULL (default) rownames of x will be applied as legend.text.}

\item{col}{Vector containing the color of bars. If NULL (default) colors are generated based on the rainbow-palette.}

\item{axes}{Logical value indicating whether to plot axes. Defaults to T.}

\item{add}{Logical value indicating whether to plot the barplot to the current device. Defaults to F.}

\item{...}{Additional graphical parameters for barplot.}
}
\description{
Custom barplot with labeled bars.
}
\details{
Plots a bar and adds labels and numbers to bars. Optionally allows for plotting standard deviations around each bar.
}
\examples{
bp(data.frame(group1=c(variable1=1,variable2=2),group2=c(variable1=3,variable2=4)),horiz=T,main="Example")
}
\keyword{plotting}
