% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.r
\name{bp}
\alias{bp}
\title{Function bp}
\usage{
bp(x = NULL, sd = NULL, cex = 1, beside = T, horiz = F,
  add.numbers = F, ndigits = 2, ncex = NA, nsrt = 0, npos = NA,
  ncol = "black", sdcol = "grey", grid = T, plot = T,
  main = NULL, sub = NULL, xlim.factor = 1.5, las = 1, srt = 45,
  xlim = NULL, ylim = NULL, names.arg = NULL, legend.text = NULL,
  optimize.legend = T, args.legend = NULL, density = NULL,
  angle = 45, col = NULL, grid.col = "grey", axes = T, add = F,
  adj = 0.5, default.labels = F, xlab = NA, ylab = NA,
  border = NA, grid.mode = 0, main1 = NULL, main2 = NULL,
  main3 = NULL, adj.main1 = 0, adj.main2 = 0, adj.main3 = 0,
  col.main1 = "black", col.main2 = "black", col.main3 = "black",
  cex.main1 = 1.2, cex.main2 = 1.2, cex.main3 = 1.2,
  font.main1 = 1, font.main2 = 2, font.main3 = 4, omitZeros = T,
  mar = NA, automar = F, ...)
}
\arguments{
\item{x}{Numeric vector, matrix or data.frame containing the values to be displayed.}

\item{sd}{Numeric vector, matrix or data.frame of same format as x containing standard deviations.}

\item{cex}{Size of axis fonts. Defaults to 1.}

\item{beside}{Logical value indicating if bars should be placed next to each other. Defaults to T.}

\item{horiz}{Logical value indicating if bars should be placed horizontal. Defaults to F.}

\item{add.numbers}{Logical value indicating if numbers should be placed above bars. Defaults to F.}

\item{ndigits}{Numeric value specifying the number of digits to be plotted (if add.numbers==T). Defaults to 2.}

\item{ncex}{Size of fonts. If NA (default) is set to cex.}

\item{nsrt}{Numeric value specifying the rotation of the numbers (between 0 and 360 degrees). Defaults to 0.}

\item{npos}{Numeric value specifying the position of numbers. If NA (default) it is determined automatically.}

\item{ncol}{Vector containing the color of bars. Defaults to "black".}

\item{sdcol}{Vector containing the color of arrows. Defaults to "grey".}

\item{grid}{Logical value indicating whether to plot a grid. Defaults to T.}

\item{plot}{Logical value indicating whether to plot the barplot. Defaults to T.}

\item{main}{Character vector with one element containing the barplot's title. Defaults to NULL}

\item{sub}{Character vector with one element containing the barplot's subtitle. Defaults to NULL}

\item{xlim.factor}{Numeric value for adding extra space to the right of the barplot (if a legend is provided). Defaults to 1.5}

\item{las}{Numeric value specifying the rotation of the y-axis (0 for 90 percent rotation, 1 for 0 percent rotation). Defaults to 1.}

\item{srt}{Numeric value specifying the rotation of the x-axis (between 0 and 360 degrees). Defaults to 45.}

\item{xlim}{Numeric vector containing xlim.}

\item{names.arg}{Character vector containing names of bars. If NULL (default) colnames of x will be applied as names.arg.}

\item{legend.text}{Legend text. Set to NA or to F to supress legend. If NULL (default) rownames of x will be applied as legend.text.}

\item{optimize.legend}{Logical or character value. If T (Default), the legend is shifted to the right if overlapping with bars (as long as no args.legend are provided). If a position is provided instead ("bottom", "bottomleft", "left","topleft","top","topright","right","bottomright") the legend will be placed and shifted accordingly. Adding a "1" at the end of a position (e.g., "top1") results in the legend being drawn horizontally instead of vertically.}

\item{args.legend}{List with arguments to pass to legend(). Defaults to list(bg = "white").}

\item{density}{Vector giving the density of shading lines, in lines per inch, for the bars or bar components. The default value of NULL means that no shading lines are drawn. Non-positive values of density also inhibit the drawing of shading lines.}

\item{angle}{Slope of shading lines, given as an angle in degrees (counter-clockwise), for the bars or bar components.}

\item{col}{Vector containing the color of bars. If NULL (default) colors are generated based on the rainbow-palette.}

\item{grid.col}{Character or rgb value containing the color of grid lines. Defaults to "grey".}

\item{axes}{Logical value indicating whether to plot axes. Defaults to T.}

\item{add}{Logical value indicating whether to plot the barplot to the current device. Defaults to F.}

\item{adj}{Numeric value between 0 and 1, indicating where to plot main title and axis labels - if any - between left (0) and right (1). Defaults to 0.5}

\item{default.labels}{Logical value, indicating whether to use traditional barplot-labels (which, e.g., hide in case of overlap) instead of bp-style labels. Defaults to F.}

\item{xlab}{Character value representing the label to be drawn next to the x-axis. Defaults to NA.}

\item{ylab}{Character value representing the label to be drawn next to the y-axis. Defaults to NA.}

\item{border}{Color of the bars' border. Defaults to NA.}

\item{grid.mode}{Numeric value specifying when to plot the grid. 0 for grid in the background, 1 for grid in the foreground. Defaults to 0.}

\item{...}{Additional graphical parameters for barplot.}
}
\description{
Custom barplot with labeled bars.
}
\details{
Plots a barplot and adds labels and numbers to bars. Optionally allows for plotting standard deviations around each bar.
}
\examples{
bp(data.frame(group1=c(variable1=1,variable2=2),group2=c(variable1=3,variable2=4)),horiz=T,main="Example")
}
\keyword{plotting}
