% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherenceLDA.r
\name{coherenceLDA}
\alias{coherenceLDA}
\title{Function coherenceLDA}
\usage{
coherenceLDA(lda, DTM, N = 0)
}
\arguments{
\item{lda}{Object of class LDA.}

\item{DTM}{document -term matrix.}
}
\description{
Calculates semantic coherence of an LDA.
}
\examples{

}
\keyword{modeling}
