% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDocs.r
\name{compareDocs}
\alias{compareDocs}
\title{Function compareDocs}
\usage{
compareDocs(
  a = NULL,
  b = NULL,
  relative = T,
  vertLine = T,
  horizLines = T,
  max = 40,
  main = "Comparison of Word Frequencies",
  pch = 16,
  cex = 0.7,
  ...
)
}
\arguments{
\item{a}{Character value.}

\item{b}{Character value.}

\item{relative}{Logical value. Defaults to T.}

\item{vertLine}{Logical value specifying whether to plot a vertical line at x=0. Defaults to T.}

\item{horizLines}{Logical value specifying whether to plot horizontal lines between each point and the y axis. Defaults to T.}

\item{max}{Numeric value indicating the maximum number of word frequencies to be displayed. Defaults to 40.}

\item{main}{Character value specifying the plot's title. Defaults to "Comparison of Word Frequencies"}

\item{pch}{Numeric element speccifying the symbol to represent points. Defaults to 16}

\item{cex}{Numeric value indicating the text size. Defaults to .7.}
}
\description{
Plots difference between two documents as a modified dotchart.
}
\details{
Plots difference between two documents as a modified dotchart. Returns matrix with relative frequencies and differences between documents a and b.
}
\examples{
compareDocs("This is document 1","this is another document")
}
\keyword{plotting}
