% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.r
\name{dotplot}
\alias{dotplot}
\title{Function dotplot}
\usage{
dotplot(x = NULL, labels = NULL, groups = NULL, gdata = NULL,
  cex = par("cex"), pt.cex = cex, pch = 16, gpch = 21,
  bg = par("bg"), col = NULL, color = par("fg"),
  gcolor = par("fg"), lcolor = "gray", xlim = range(x[is.finite(x)]),
  main = NULL, xlab = NULL, ylab = NULL, deviations = NA,
  x2 = NA, vertLine = 0, horizLines = T, max = 40, sort = F,
  add.numbers = F, ndigits = 2, ncex = NA, nsrt = 0, npos = NA,
  ncol = "black", main1 = NULL, main2 = NULL, main3 = NULL,
  adj.main1 = 0, adj.main2 = 0, adj.main3 = 0, col.main1 = "black",
  col.main2 = "black", col.main3 = "black", cex.main1 = 1.2,
  cex.main2 = 1.2, cex.main3 = 1.2, font.main1 = 1, font.main2 = 2,
  font.main3 = 4, omitZeros = F, xaxs = "i", yaxs = "i", ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{cex}{Numeric value indicating the text size. Defaults to .7.}

\item{pch}{Numeric element speccifying the symbol to represent points. Defaults to 16}

\item{main}{Character value specifying the plot's title. Defaults to "Comparison of Word Frequencies"}

\item{vertLine}{Logical value specifying whether to plot a vertical line at x=0. Defaults to T.}

\item{horizLines}{Logical value specifying whether to plot horizontal lines between each point and the y axis. Defaults to T.}

\item{max}{Numeric value indicating the maximum number of frequencies to be displayed. Defaults to 40.}

\item{add.numbers}{Logical value indicating if numbers should be placed next to the dots. Defaults to F.}

\item{ndigits}{Numeric value specifying the number of digits to be plotted (if add.numbers==T). Defaults to 2.}

\item{ncex}{Size of fonts. If NA (default) is set to cex.}

\item{nsrt}{Numeric value specifying the rotation of the numbers (between 0 and 360 degrees). Defaults to 0.}

\item{npos}{Numeric value specifying the position of numbers. If NA (default) it is determined automatically.}

\item{ncol}{Vector containing the color of bars. Defaults to "black".}
}
\description{
Plots a modified dotchart.
}
\details{
Plots vector as modified dotchart.
}
\examples{
dotplot(rnorm(100))
}
\keyword{plotting}
