% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTable.r
\name{extractTable}
\alias{extractTable}
\title{Function extractTable}
\usage{
extractTable(
  x,
  reg.up = NULL,
  reg.down = NULL,
  reg.left = "^",
  reg.right = "$",
  reg.fix = NULL,
  header = F,
  trim = T,
  convert = F,
  correctNotation = F
)
}
\arguments{
\item{x}{A character element containing all the text toextract the table from (or a character vector containing one text-line per element).}

\item{reg.up}{Regular expression spefifying the top of the table. Defaults to NULL.}

\item{reg.down}{Regular expression spefifying the bottom of the table. Defaults to NULL.}

\item{reg.left}{Regular expression spefifying the bottom of the table. Defaults to "^".}

\item{reg.right}{Regular expression spefifying the bottom of the table. Defaults to "$"}

\item{reg.fix}{Regular expression spefifying the bottom of the table. Defaults to NULL.}

\item{header}{Logical value that specifies, whether to treat first row and column as row- and column-names. Defaults to F.}

\item{trim}{Logical value that specifies, whether to trim entries. Defaults to T.}

\item{convert}{Logical value that specifies, whether to convert table to numeric (after correcting for notation in case correctNotation is set to T). Please use with care, if table contains non-numeric information. Defaults to T.}

\item{correctNotation}{Logical value that specifies, whether to correct for the fact that entries use commata instead of dots. Erases all dots in all entries and subsequently replaces commata by dots. Defaults to F.}
}
\description{
Extracts table as data.frames from multiple lines of text.
}
\details{
Extracts table as data.frames from multiple lines of text. Columns are assumed to be separated by whitespaces that are placed at the same position in each line. Please make sure that headers don't span multiple columns or rows.
}
\keyword{scraping}
