% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHTML.r
\name{getHTML}
\alias{getHTML}
\title{Function getHTML}
\usage{
getHTML(
  url = "https://scholar.google.de/citations?user=-TjY7oEAAAAJ&hl=de&oi=sra",
  encoding = "UTF-8",
  save = F,
  filename = NULL,
  closeCon = F,
  method = "libcurl",
  silent = T,
  curlGetHeaders = F,
  browseURL = F,
  ...
)
}
\arguments{
\item{url}{Character string speficying a valid URL.}

\item{encoding}{Character value specifying the encoding. Defaults to "UTF-8".}

\item{save}{Logical value specifying whether the HTML-code should be saved to a txt-file.}

\item{filename}{Character value specifying the filename (if save is TRUE). If NULL (default) as.numeric(Sys.time()) is applied.}

\item{closeCon}{Logical value specifying whether to close open connections before and after scraping. defaults to F.}

\item{method}{Character value specifying the method. Defaults to "libcurl".}

\item{silent}{Logical value specifying whether to skip informative messages. Defaults to T.}

\item{browseURL}{Locgical value specifying wether to open the URL in a web-browser. Defaults to F.}

\item{curlHeaders}{Logical value specifying whether to apply curlHeaders. Defaults to F.}

\item{headers}{Logical value.}
}
\description{
Returns HTML from an URL obained via GET-request.
}
\details{
Returns HTML from an URL obained via GET-request. Assumes UTF-8 encoding by default. Returns a character element containing HTML-code.
}
\examples{
getHTML()
}
\keyword{scraping}
