% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearchLDA.r
\name{gridSearchLDA}
\alias{gridSearchLDA}
\title{Function gridSearchLDA}
\usage{
gridSearchLDA(dtm, k = NULL, alpha = NULL, delta = NULL,
  grid = NULL, fun = function(x) -1 * AIC(x), best = T, plot = T,
  ...)
}
\description{
Function gridSearchLDA
}
\keyword{modeling}
