% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXY.r
\name{plotXY}
\alias{plotXY}
\title{Function plotXY}
\usage{
plotXY(x = NULL, y = NULL, na.rm = T, color1 = rgb(0/255, 84/255,
  122/255, 0.7), color2 = "darkorange", show.text = T, xlab = "x",
  ylab = "y", main = "Correlation", pch = 16, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector of the same length as x.}

\item{na.rm}{Logical value indicating whether missing values should be skipped. Defaults to T.}

\item{color1}{Color of points in the scattergram. Defaults to rgb(0/255,84/255,122/255,.7).}

\item{color2}{Color of the regression line. Defaults to "darkorange".}

\item{show.text}{Logical value indicating whether text should be added to the regression line. Defaults to T.}
}
\description{
Plots bivariate correlation based on two numeric vectors.
}
\details{
Plots scattergram and bivariate correlation based on two numeric vectors.
}
\examples{
plotXY()
}
\keyword{plotting}
