% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXY.r
\name{plotXY}
\alias{plotXY}
\title{Function plotXY}
\usage{
plotXY(
  x = NULL,
  y = NULL,
  complexity = 0,
  rep.nnet = 10,
  attrModel = T,
  na.rm = T,
  color1 = rgb(0, 0, 0, 0.7),
  color2 = rgb(0, 0, 1),
  color3 = rgb(0, 0, 1, 0.2),
  xlab = "x",
  ylab = "y",
  axes = T,
  add = F,
  main = NA,
  sub = NA,
  pch = 16,
  lwd = 2,
  cex = 0.7,
  cex.sub = 0.7,
  generalize = F,
  main1 = NULL,
  main2 = NULL,
  main3 = NULL,
  mar = NA,
  adj.main1 = 0,
  adj.main2 = 0,
  adj.main3 = 0,
  col.main1 = "black",
  col.main2 = "black",
  col.main3 = "black",
  cex.main1 = 1.2,
  cex.main2 = 1.2,
  cex.main3 = 1.2,
  font.main1 = 1,
  font.main2 = 2,
  font.main3 = 4,
  ...
)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector of the same length as x.}

\item{complexity}{Numeric value specifying the amount of nonlinearity modelled. Defaults to 0 (i.e., a linear model).}

\item{rep.nnet}{Numeric value specifying the number of nnet-objects to choose the best model from.}

\item{attrModel}{Logical value specifying whether to add the model as an attribute to the object returned.}

\item{na.rm}{Logical value indicating whether missing values should be skipped. Defaults to T.}

\item{color1}{Color of points in the scattergram. Defaults to rgb(0,0,0,.7).}

\item{color2}{Color of the regression line. Defaults to rgb(0,0,1).}

\item{color3}{Color of the prediction interval. Defaults to rgb(0,0,1,.2).}

\item{...}{additional parameters passed to the plot function.}
}
\description{
Plots bivariate correlation based on two numeric vectors.
}
\details{
Plots scattergram and bivariate correlation based on two numeric vectors.
}
\examples{
plotXY()
}
\keyword{plotting}
