% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDF.r
\name{reduceDF}
\alias{reduceDF}
\title{Function reduceDF}
\usage{
reduceDF(data, output = NULL, cutoff = 0.1,
  control = caret::trainControl(method = "boot", number = 30))
}
\arguments{
\item{data}{data.frame containing data.}

\item{output}{index or name of criterion in data.}

\item{y}{numeric vector.}
}
\description{
Applies LASSO regression and returns a data.frame with the most important predictors.
}
\examples{
data=data.frame(y=rnorm(100),x1=rnorm(100),x2=rnorm(100));
reduceDF(data,"y1")
}
\keyword{preprocessing}
