% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeJDDM.r
\name{scrapeJDDM}
\alias{scrapeJDDM}
\title{Function scrapeJDDM}
\usage{
scrapeJDDM(plot = T)
}
\arguments{
\item{plot}{Logical vakue specifying wether to plot data on reads and downloads.}
}
\description{
Downloads and returns data and metadata from the database of the Journal of Dynamic Decision Making.
}
\details{
Downloads and returns data and metadata from the database of the Journal of Dynamic Decision Making.
}
\examples{
s=scrapeJDDM();
}
\keyword{scraping}
