% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeJDDM.r
\name{scrapeJDDM}
\alias{scrapeJDDM}
\title{Function scrapeJDDM}
\usage{
scrapeJDDM(plot = T,
  urls = paste0("https://journals.ub.uni-heidelberg.de/index.php/jddm/issue/view/",
  c(3269, 2836, 2269, 3694, 4432)),
  url2 = "https://journals.ub.uni-heidelberg.de/cgi-bin/oastats.cgi?repo=ojs;from_date=2015-09-29\%2021:39:27;id=jddm:")
}
\arguments{
\item{plot}{Logical value specifying wether to plot data on reads and downloads.}

\item{urls}{Character vector specifying the URLs of the individual issues to scrape.}

\item{url2}{Character value specifying the URL of global journal statistics.}

\item{plot}{Logical value specifying wether to plot data on reads and downloads.}
}
\description{
Downloads and returns data and metadata from the database of the Journal of Dynamic Decision Making.
}
\details{
Downloads and returns data and metadata from the database of the Journal of Dynamic Decision Making.
}
\examples{
s=scrapeJDDM();
}
\keyword{scraping}
