% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderplot.r
\name{spiderplot}
\alias{spiderplot}
\title{Function spiderplot}
\usage{
spiderplot(x = NULL, lower = NULL, upper = NULL, weights = NULL,
  names = NULL, main = NULL, max = 1, xylim = 1.5, add.scale = T)
}
\arguments{
\item{x}{Numeric vector containing the values to be displayed. Values should be between 0 and 1.}

\item{lower}{Numeric vector of the same length as x, containing the values to be displayed as an lower bound. Not displayed if NULL (if !is.null(x)).}

\item{upper}{Numeric vector of the same length as x, containing the values to be displayed as an upper bound. Not displayed if NULL (if !is.null(x)).}

\item{weights}{Numeric vector of the same length as x, containing the relative length of dimensions. Values should be between 0 and 1.}

\item{main}{Character vector with one element containing the barplot's title. Defaults to NULL}

\item{max}{Numeric value determining the largest possible value of the data to be displayed. Defaults to 1.}

\item{xylim}{Numeric value determining both upper and lower limits on both x and y axis (xlim & ylim). Defaults to 1.5.}
}
\description{
Plots a radar chart based on a numeric vector.
}
\details{
Plots a radar chart based on a numeric vector. Each dimension of the radar chart can be assigned a weight and/or an upper and lower bound.
}
\examples{
spiderplot()
}
\keyword{plotting}
