% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderplot.r
\name{spiderplot}
\alias{spiderplot}
\title{Function spiderplot}
\usage{
spiderplot(x = NULL, lower = NULL, upper = NULL, weights = NULL,
  names = NULL, main = NULL, max = 1, min = 0, xylim = 1.5,
  add.scale = T, col = "#00547A", col2 = "#FFA500E6", border = NA,
  mode = 0, arrows.lwd = 3, arrows.length = 0.01, add = F,
  add.numbers = F, numbers = 2, add.grid = T, add.labels = T,
  line.type = "o", grid.detail = 1, grid.points = NULL)
}
\arguments{
\item{x}{Numeric vector containing the values to be displayed. Values should be between 0 and 1.}

\item{lower}{Numeric vector of the same length as x, containing the values to be displayed as an lower bound. Not displayed if NULL (if !is.null(x)).}

\item{upper}{Numeric vector of the same length as x, containing the values to be displayed as an upper bound. Not displayed if NULL (if !is.null(x)).}

\item{weights}{Numeric vector of the same length as x, containing the relative length of dimensions. Values should be between 0 and 1.}

\item{main}{Character vector with one element containing the barplot's title. Defaults to NULL}

\item{max}{Numeric value determining the largest possible value of the data to be displayed. Defaults to 1.}

\item{xylim}{Numeric value determining both upper and lower limits on both x and y axis (xlim & ylim). Defaults to 1.5.}

\item{col}{Character or rgb value specifying the line's color. Defaults to rgb(0,0,1).}

\item{col2}{Character or rgb value specifying the color of the area between upper and lower bound. Defaults to rgb(0,0,1,.5).}

\item{border}{Character or rgb value specifying the color of the polygon border. Defaults to NA.}

\item{mode}{Numeric value specifying whether to draw a polygon (mode=1) or arrows (mode!=1), and whether to draw a polygon for x (mode>=0) or not (mode<0). Defaults to 0.}

\item{arrows.lwd}{Numeric value specifying lwd for arrows (if any). Defaults to 2.}

\item{arrows.length}{Numeric value specifying lwd for arrows (if any). Defaults to .01.}

\item{add}{Logical value specifying whether to add graph to the current device. Defaults to F.}

\item{add.numbers}{Logical value specifying whether to add numbers to labels. Defaults to F.}

\item{add.grid}{Logical value specifying whether to add numbers to add grid lines. Defaults to T.}

\item{add.labels}{Logical value specifying whether to add labels to the plot. Defaults to T.}
}
\description{
Plots a radar chart based on a numeric vector.
}
\details{
Plots a radar chart based on a numeric vector. Each dimension of the radar chart can be assigned a weight and/or an upper and lower bound.
}
\examples{
spiderplot()
}
\keyword{plotting}
