% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetHTML.r
\name{subsetHTML}
\alias{subsetHTML}
\title{Function subsetHTML}
\usage{
subsetHTML(html, tag = "div", pattern = NULL, edit = F, save = F,
  plot = F, prefix = NULL, trim = T)
}
\arguments{
\item{html}{A character element containing HTML-code.}

\item{tag}{Character element specifying the the subsets of interest. Defaults to "div".}

\item{pattern}{Regular expression further specifying subsets of interest. If NULL (default) equals tag.}

\item{edit}{Logical value specifying whether the data.frame should be plotted/edited.}

\item{save}{Logical value specifying whether the HTML-code should be saved to a csv-file.}

\item{prefix}{Character value specifying the beginning of the filename (in case of saving). If NULL (default) as.numeric(Sys.time()) is applied.}

\item{trim}{Logical value specifying whether to trim text. Defaults to T.}
}
\description{
Extracts a coherent subset of code from HTML-code.
}
\details{
Extracts a coherent subset of code from HTML-code.
}
\examples{
subsetHTML(getHTML("https://jobs.meinestadt.de/nuernberg/suche?words=Wissenschaftlicher\%20Mitarbeiter",tag="div",pattern="class=\\"m-resultListEntries__content\\""))
}
\keyword{scraping}
