% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeLDA.r
\name{summarizeLDA}
\alias{summarizeLDA}
\title{Function summarizeLDA}
\usage{
summarizeLDA(lda, data, dtm, main = "Ergebnisse", stopwords = NULL)
}
\arguments{
\item{lda}{Object of class LDA.}

\item{data}{character vector of documents.}

\item{dtm}{document-term matrix}
}
\description{
Topic summary of LDA.
}
\examples{
 
}
\keyword{modeling}
