% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeSTM.r
\name{summarizeSTM}
\alias{summarizeSTM}
\title{Function summarizeSTM}
\usage{
summarizeSTM(stm, data, topicNo = 0, main = "Results",
  stopwords = NULL, cex = 1, simlpe = F)
}
\arguments{
\item{stm}{Object of class stm.}

\item{data}{character vector of documents.}
}
\description{
Topic summary of stm.
}
\examples{

}
\keyword{modeling}
