% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textbreaker.r
\name{textbreaker}
\alias{textbreaker}
\title{Function textbreaker}
\usage{

  textbreaker(text = "Let's write some text with many different words and plot it in a nice little device.",
  maxlength = 30, maxlines = 10, centered = F, separator = "\\n",
  trim = T, plot = F, ...)
}
\arguments{
\item{text}{Character value containing the text to be split into multiple lines.}

\item{maxlength}{Numeric value specifying the maximum number of characters per line.}
}
\description{
Adds linebreaks to text.
}
\details{
Adds linebreaks to text.
}
\examples{
textbreaker("hello world hello world hello world hello world")
}
\keyword{helper}
