% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfIdf.r
\name{tfIdf}
\alias{tfIdf}
\title{Function tfIdf}
\usage{
tfIdf(dtm, docs = NA, relative = T, bool = F)
}
\arguments{
\item{dtm}{Document Term Matrix.}

\item{docs}{Numeric element specifying a document in the dtm. If NA (default tfIdf is calculated for each document)}
}
\description{
Term-Frequency-Inverse-Term-Frequency.
}
\keyword{mining}
\keyword{text}
